webclipper(build): move entrypoints under src
Configure WXT entrypointsDir to src/entrypoints and move existing entrypoints there.

Update entrypoint import paths, Tailwind config, AMO source packager, and AGENTS docs.

Verified: npm --prefix Extensions/WebClipper run compile; npm --prefix Extensions/WebClipper run test; npm --prefix Extensions/WebClipper run build
