+++ Extensions/WebClipper/AGENTS.md
@@ -65,36 +65,36 @@
 
 ### 模块入口索引（2026-03，WXT 新架构）
 
-- **WXT Background 入口**：`entrypoints/background.ts`
-- **WXT Content 入口**：`entrypoints/content.ts`
-- **WXT Popup 入口**：`entrypoints/popup/*`
-- **WXT App 入口**：`entrypoints/app/*`
+- **WXT Background 入口**：`src/entrypoints/background.ts`
+- **WXT Content 入口**：`src/entrypoints/content.ts`
+- **WXT Popup 入口**：`src/entrypoints/popup/*`
+- **WXT App 入口**：`src/entrypoints/app/*`
 - **消息协议（前后端共享）**：`src/platform/messaging/message-contracts.ts`
   - 统一 `CORE_MESSAGE_TYPES` / `NOTION_MESSAGE_TYPES` / `OBSIDIAN_MESSAGE_TYPES` / `UI_MESSAGE_TYPES`，禁止在 popup/background 中散落硬编码 type 字符串。
 - **WebExtensions API 封装层**：`src/platform/webext/`
   - `tabs.ts` / `scripting.ts` / `web-navigation.ts` 统一封装 `chrome.*` 与 `browser.*` 差异；业务模块优先调用封装函数，避免散落全局 API 访问。
 - **兼容层已移除**：`src/runtime-context.ts` + `src/export/bootstrap.ts` 已删除；禁止依赖 `globalThis.WebClipper` 的隐式注入。
-- **后台初始化与路由（当前）**：`entrypoints/background.ts` + `src/bootstrap/background-services.ts` + `src/platform/messaging/background-router.ts`
-  - `entrypoints/background.ts` 负责创建 services、注册各类 handlers，并启动 router。
+- **后台初始化与路由（当前）**：`src/entrypoints/background.ts` + `src/bootstrap/background-services.ts` + `src/platform/messaging/background-router.ts`
+  - `src/entrypoints/background.ts` 负责创建 services、注册各类 handlers，并启动 router。
 - **本地内容库（conversations/messages）**：`src/conversations/`
   - 按职责拆分：`domain/`（类型/纯函数）+ `data/`（IndexedDB）+ `background/`（handlers/storage）+ `client/`（UI 调用）+ `content/`（增量 diff）
 - **Notion 同步模块**：`src/sync/notion/`
   - 编排入口：`src/sync/notion/notion-sync-orchestrator.ts`（由 `src/sync/background-handlers.ts` 路由触发）。
 - **Obsidian 模块**：Local REST API Sync（平台主导）
   - 已重构为 Local REST API 平台主导同步：
-  - popup：`entrypoints/popup/tabs/SettingsTab.tsx`（配置与连通性测试）+ `entrypoints/popup/tabs/ChatsTab.tsx`（触发同步）
+  - popup：`src/ui/popup/tabs/SettingsTab.tsx`（配置与连通性测试）+ `src/ui/popup/tabs/ChatsTab.tsx`（触发同步）
   - background：`src/sync/obsidian/obsidian-sync-orchestrator.ts`。
 - **Web Article Fetch（手动抓取当前页）**：`src/collectors/web/article-fetch.ts` + `src/collectors/web/article-fetch-background-handlers.ts`
   - background 侧通过 `scriptingExecuteScript`（`src/platform/webext/scripting.ts`）注入 `src/vendor/readability.js` 并抽取正文，写入本地 conversations/messages（kind=article）。
 - **UI：消息气泡与 Markdown 渲染（共享）**：`src/ui/shared/ChatMessageBubble.tsx` + `src/ui/shared/markdown.ts`
   - popup 与 app 共用同一套 bubble + renderer，避免“同一条消息在两处渲染不一致”。
-- **Inpage 显示范围设置**：`entrypoints/content.ts` + `src/bootstrap/content.ts`
-  - popup 写入：`entrypoints/popup/tabs/SettingsTab.tsx`（保存 `inpage_supported_only` 到 `chrome.storage.local`）。
+- **Inpage 显示范围设置**：`src/entrypoints/content.ts` + `src/bootstrap/content.ts`
+  - popup 写入：`src/ui/popup/tabs/SettingsTab.tsx`（保存 `inpage_supported_only` 到 `chrome.storage.local`）。
   - content script 统一匹配所有 `http(s)` 页面，在运行时读取 `inpage_supported_only` 决定是否启动 inpage controller（避免依赖动态注册 content scripts 的浏览器兼容差异）。
   - 当前实现仅在启动时读取配置，因此切换开关后需刷新页面生效。
 
 Phase 3（JS→TS）收口状态：
-- `src + entrypoints` 运行时代码已收敛为 TS 主实现。
+- `src` 运行时代码已收敛为 TS 主实现（入口文件也已迁入 `src/entrypoints/`）。
 - runtime JS allowlist 仅保留第三方脚本资产：`public/src/vendor/readability.js`（产物路径仍为 `src/vendor/readability.js`）。
 
 ### Obsidian 约束
