+++ Extensions/WebClipper/src/entrypoints/app/index.html
