+++ Extensions/WebClipper/src/entrypoints/app/main.tsx
@@ -1,8 +1,8 @@
 import React from 'react';
 import ReactDOM from 'react-dom/client';
-import AppShell from '../../src/ui/app/AppShell';
-import '../../src/ui/styles/tailwind.css';
-import '../../src/ui/styles/tokens.css';
+import AppShell from '../../ui/app/AppShell';
+import '../../ui/styles/tailwind.css';
+import '../../ui/styles/tokens.css';
 import './style.css';
 
 ReactDOM.createRoot(document.getElementById('root')!).render(
