+++ Extensions/WebClipper/src/entrypoints/popup/main.tsx
@@ -1,8 +1,8 @@
 import React from 'react';
 import ReactDOM from 'react-dom/client';
-import PopupShell from '../../src/ui/popup/PopupShell';
-import '../../src/ui/styles/tailwind.css';
-import '../../src/ui/styles/tokens.css';
+import PopupShell from '../../ui/popup/PopupShell';
+import '../../ui/styles/tailwind.css';
+import '../../ui/styles/tokens.css';
 import './style.css';
 
 ReactDOM.createRoot(document.getElementById('root')!).render(
