+++ Extensions/WebClipper/src/entrypoints/popup/style.css
